+++ src/app/App.svelte
@@ -8,7 +8,7 @@
   import {identity, isNil, last} from "ramda"
   import {first} from "hurdak/lib/hurdak"
   import {warn} from "src/util/logger"
-  import {timedelta, hexToBech32, bech32ToHex, shuffle, now} from "src/util/misc"
+  import {timedelta, hexToBech32, bech32ToHex, shuffle, now, tryFunc} from "src/util/misc"
   import cmd from "src/agent/cmd"
   import {onReady, relays} from "src/agent/db"
   import keys from "src/agent/keys"
@@ -38,6 +38,8 @@
   document.head.append(style)
 
   $: style.textContent = `:root { ${getThemeVariables($theme)}; background: var(--gray-8); }`
+
+  tryFunc(() => navigator.registerProtocolHandler("web+nostr", `${location.origin}/%s`))
 
   const seenChallenges = new Set()