+++ src/app/views/NoteShare.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+  import {nip19} from "nostr-tools"
+  import Content from "src/partials/Content.svelte"
+  import QRCode from "src/partials/QRCode.svelte"
+  import {getRelayForEventHint} from "src/agent/relays"
+
+  export let note
+
+  const nevent = nip19.neventEncode({id: note.id, relays: [getRelayForEventHint(note)]})
+</script>
+
+<Content size="lg">
+  <QRCode code={`nostr:${nevent}`} />
+  <div class="text-center text-gray-1">Copy or scan from a nostr app to share this note.</div>
+</Content>